Add files via upload
Attrition:- company losing its customer base
Attrition is a process in which the workforce dwindles at a company, following a period in which a number of people retire or resign, and are not replaced.

A reduction in staff due to attrition is often called a hiring freeze and is seen as a less disruptive way to trim the workforce and reduce payroll than layoffs
In this NoteBook our Aim will be to analyze the datasets completely wrt each and feature and find the reasin behind Attrition of Employees.
And what the top factors which lead to employee attrition?
Description about the data
Education
1 'Below College' 2 'College' 3 'Bachelor' 4 'Master' 5 'Doctor'

EnvironmentSatisfaction
1 'Low' 2 'Medium' 3 'High' 4 'Very High'

JobInvolvement
1 'Low' 2 'Medium' 3 'High' 4 'Very High'

JobSatisfaction
1 'Low' 2 'Medium' 3 'High' 4 'Very High'

PerformanceRating
1 'Low' 2 'Good' 3 'Excellent' 4 'Outstanding'

RelationshipSatisfaction
1 'Low' 2 'Medium' 3 'High' 4 'Very High'

WorkLifeBalance
1 'Bad' 2 'Good' 3 'Better' 4 'Best'

Additinal Description Which will help to understand the features
Hourly Rate: An hourly employee is paid a wage for each hour worked, unlike a salaried employee.

EmployeeNumber: An Employee Number is a unique number that has been assigned to each current and former State employee and elected official in the Position and Personnel DataBase (PPDB).

JobLevel: Job levels, also known as job grades and classifications, set the responsibility level and expectations of roles at your organization. They may be further defined by impact,
          seniority, knowledge, skills, or job title, and are often associated with a pay band. The way you structure your job levels should be dictated by the needs of your unique organization and teams.

StockOptionLevel: Employee stock options, also known as ESOs, are stock options in the company s stock granted by an employer to certain employees. 
                  Typically they are granted to those in management or officer-level positions. Stock options give the employee the right to buy a certain amount of stock at a specific price, during a specific period of time. 
                  Options typically have expiration dates as well, by which the options must have been exercised, otherwise they will become worthless.